canvas2d/path-objects/2d.path.arcTo.negative.html

<!DOCTYPE html>
<title>Canvas test: 2d.path.arcTo.negative</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.path.html">path</a>.<a href="index.2d.path.arcTo.html">arcTo</a>.negative</h1>
<p class="desc">arcTo() with negative radius throws an exception</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.path.arcTo.negative">2d.path.arcTo.negative</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("arcTo() with negative radius throws an exception");
_addTest(function(canvas, ctx) {

assert_throws("INDEX_SIZE_ERR", function() { ctx.arcTo(0, 0, 0, 0, -1); });


});
</script>
